backend/requirements.txt
flask==0.12.2
python-dotenv==0.7.1
watson-developer-cloud==1.0.2
flask-cors==3.0.3
